+++ packages/cmsui/components/Toolbar/Toolbar.tsx
@@ -14,7 +14,7 @@ const Toolbar = () => {
       id="toolbar"
       className="bg-quanta-smoke flex h-full w-[80px] justify-center transition-[left,right,width] duration-200 ease-linear"
     >
-      <div className="fixed inset-y-0 z-10 my-4 flex flex-col items-center justify-between">
+      <div className="fixed inset-y-0 z-10 my-4 flex flex-col items-center gap-4">
         <Pluggable name="toolbar-top" />
         <Pluggable name="toolbar-bottom" />
         <Plug pluggable="toolbar-bottom" id="button-settings">
